Adding unem for data storage operation
Adding unem for data storage operation instead of using boolean for status checking.
